Fix Windows build after r153134. https://bugs.webkit.org/show_bug.cgi?id=119090
Reviewed by Oliver Hunt.

Add USE(PTHREADS) guards around pthread specific code
for now. This will cause isCompilationThread to always
return false on non pthread platforms such as Windows.

We might be able to use the Windows one-time initialization
for this, but this is only available from Windows Vista on.

* wtf/CompilationThread.cpp:
(WTF::initializeCompilationThreads):


Identifier: 137120@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@153336 268f45cc-cd09-0410-ab3c-d52691b4dbfc