Speculative fix for crash due to string access on multiple threads https://bugs.webkit.org/show_bug.cgi?id=119279 <rdar://problem/14267833>
Reviewed by Darin Adler.

We can't use String::isolatedCopy for passing the local storage directory to the storage thread since
that returns a String that's copied and then destroyed after the call to bind returns, leaving a small window
where the refcount can be accessed simultaneously from two threads.

Work around this by passing a PassRefPtr<StringImpl> to bind instead; the act of copying the PassRefPtr will
clear out the original and so when the original is destroyed the underlying StringImpl pointer will be null.

* UIProcess/Storage/LocalStorageDatabaseTracker.cpp:
(WebKit::LocalStorageDatabaseTracker::setLocalStorageDirectory):
(WebKit::LocalStorageDatabaseTracker::setLocalStorageDirectoryInternal):
* UIProcess/Storage/LocalStorageDatabaseTracker.h:

Identifier: 137262@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@153499 268f45cc-cd09-0410-ab3c-d52691b4dbfc