Layout should force a StyleResolver rebuild if there isn't one at all. <http://webkit.org/b/119378>
Reviewed by Antti Koivisto.

We can't tell if viewport-dependent media queries have been invalidated at layout,
unless we have a StyleResolver to ask about this.
In case there isn't one, take the same path as we would with an invalidated query
instead of relying on ensureStyleResolver() to build it.

* page/FrameView.cpp:
(WebCore::FrameView::layout):


Identifier: 137349@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@153595 268f45cc-cd09-0410-ab3c-d52691b4dbfc