ASSERT(m_type == toSVGTransform.type()) https://bugs.webkit.org/show_bug.cgi?id=113217
Reviewed by Darin Adler.

Source/WebCore:

For to animations clear the transform list while animating and reset to it after the animation is done.
This matches Firefox and Opera 12 behavior.

Tests: svg/animations/animateTransform-to-scale-expected.svg
       svg/animations/animateTransform-to-scale.svg

* svg/SVGAnimatedTransformList.cpp:
(WebCore::SVGAnimatedTransformListAnimator::calculateAnimatedValue):

LayoutTests:

Add reftest for additive to animation where from and to transform types differ.

* svg/animations/animateTransform-to-scale-expected.svg: Added.
* svg/animations/animateTransform-to-scale.svg: Added.

Identifier: 137582@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@153886 268f45cc-cd09-0410-ab3c-d52691b4dbfc